backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.32
httpx>=0.27
openai>=1.54
osmnx>=2.0
networkx>=3.3
geopandas>=0.14
shapely>=2.0
sqlalchemy>=2.0
geoalchemy2>=0.15
psycopg2-binary>=2.9
pandas>=2.2
numpy>=1.26
python-dotenv>=1.0
pydantic>=2.7
scikit-learn>=1.4

